% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmd.haircut.check.alignment}
\alias{cmd.haircut.check.alignment}
\title{Check the raw/cut/ref alignment}
\usage{
cmd.haircut.check.alignment(indir, outdir, batch.n = NA, batch.id = NA,
  prog = PR.HAIRCUT.CHCKAL)
}
\description{
Check the raw/cut/ref alignment
}
\examples{
#
#	run from command line 
#	this produces a command line string that can be run in UNIX alikes
#
\dontrun{
	
#DATA		<- SET THIS DIRECTORY
ind			<- paste(DATA, 'contigs_150408_wref', sep='/' )
cat(cmd.haircut.check.alignment(indir, indir))
}
}

