% Generated by roxygen2 (4.0.2): do not edit by hand
\name{haircut.get.training.contigs}
\alias{haircut.get.training.contigs}
\title{Generate curated contigs that are used for training}
\usage{
haircut.get.training.contigs(indir = NA, outfile = NA, ctrain = NULL)
}
\description{
Generate curated contigs that are used for training
}

