% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmd.haircut.pipeline}
\alias{cmd.haircut.pipeline}
\title{Command line generator to run the Haircut pipeline}
\usage{
cmd.haircut.pipeline(indir, outdir, batch.n = NA, batch.id = NA)
}
\description{
The haircut pipeline involves two steps.
First, descriptive statistics are calculated for each contig and the consensus sequence of that associated references.
Second, these descriptive statistics are used to calculate a call probability for 10 base pair long contig chunks.
The call probability is modeled as a function of the descriptive statistics.
The underlying statistical model is pre-computed and supplied with the R package.
Based on the call probabilites, 10bp chunks of each contig are called (yes=1, no=0).
If cut/raw contigs correspond to each other, only one of both is returned.
}
\examples{
#
#	create multiple runs on HPC using the command line version
#
\dontrun{
	
#DATA		<- SET THIS DIRECTORY
indir		<- paste(DATA, 'contigs_150408_wref', sep='/' )
outdir		<- paste(DATA, 'contigs_150408_wref_cutstat', sep='/' )		
batch.n		<- 200
tmp			<- data.table(FILE=list.files(indir, pattern='fasta$', recursive=T))
tmp[, BATCH:= ceiling(seq_len(nrow(tmp))/batch.n)]
tmp			<- tmp[, max(BATCH)]
for(batch.id in seq.int(1,tmp))
{	
	
	cmd			<- cmd.haircut.pipeline(indir, outdir, batch.n=batch.n, batch.id=batch.id)
	cmd			<- cmd.hpcwrapper(cmd, hpc.nproc= 1, hpc.q='pqeelab', hpc.walltime=4, hpc.mem="5000mb")
	cat(cmd)		
	outdir		<- paste(DATA,"tmp",sep='/')
	outfile		<- paste("hrct",paste(strsplit(date(),split=' ')[[1]],collapse='_',sep=''),sep='.')
	cmd.hpccaller(outdir, outfile, cmd)	
}	
}
}

