% Generated by roxygen2 (4.0.2): do not edit by hand
\name{haircutwrap.get.cut.statistics}
\alias{haircutwrap.get.cut.statistics}
\title{Compute descriptive statistics that are used to calculate the call probability}
\usage{
haircutwrap.get.cut.statistics(indir, par, outdir = indir, batch.n = NA,
  batch.id = NA)
}
\description{
Compute descriptive statistics that are used to calculate the call probability
}
\examples{
#
#	within R
#
\dontrun{
#DATA	<- SET THIS DIRECTORY
indir	<- paste(DATA, 'contigs_150408_wref', sep='/' )
outdir	<- paste(DATA, 'contigs_150408_wref_cutstat', sep='/' )		
par		<- c('FRQx.quantile'=NA, 'FRQx.thr'=NA, 'CNS_FRQ.window'=200, 'CNS_AGR.window'=200, 'GPS.window'=200)
haircutwrap.get.cut.statistics(indir, par, outdir=outdir)	
}
}

